% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.R
\name{G.prep}
\alias{G.prep}
\title{Variant Set Mixed Model Association Tests (StocSum)}
\usage{
G.prep(
  meta.files.prefix,
  n.files = rep(1, length(meta.files.prefix)),
  cohort.group.idx = NULL,
  group.file,
  group.file.sep = "\\t",
  auto.flip = FALSE
)
}
\arguments{
\item{n.files}{An integer vector with the same length as meta.files.prefix, indicating how many sets of intermediate files (.score.* and .var.*) are expected from each cohort, usually as the result of multi-threading in creating the intermediate files (default = rep(1, length(meta.files.prefix))).}

\item{cohort.group.idx}{A vector with the same length as meta.files.prefix, indicating which cohorts should be grouped together in the meta-analysis assuming homogeneous genetic effects. For example, c("a","b","a","a","b") means cohorts 1, 3, 4 are assumed to have homogeneous genetic effects, and cohorts 2, 5 are in another group with homogeneous genetic effects (but possibly heterogeneous with group "a"). If NULL, all cohorts are in the same group (default = NULL).}

\item{group.file}{A plain text file with 6 columns defining the test units. There should be no headers in the file, and the columns are group name, chromosome, position, reference allele, alternative allele and weight, respectively.}

\item{group.file.sep}{The delimiter in group.file (default = "\\t").}

\item{auto.flip}{A logical switch for whether to enable automatic allele flipping if a variant with alleles ref/alt is not found at a position, but a variant at the same position with alleles alt/ref is found (default = FALSE). Use with caution for whole genome sequence data, as both ref/alt and alt/ref variants at the same position are not uncommon, and they are likely two different variants, rather than allele flipping.}

\item{meta.file.prefix}{A character vector for prefix of intermediate files (*.sample.* and *.resample.*).}
}
\value{
\code{G.prep} returns a list of class G.prep with the following components:
\item{meta.file.prefix}{a character vector for prefix of intermediate files (*.sample.* and *.resample.*).}
\item{n.cohort}{the number of cohorts.}
\item{cohort.group.idx}{inherited from input argument \code{cohort.group.idx}.}
\item{n.cohort.groups}{the number of groups of cohorts.}
\item{n.groups}{the test unit group number.}
\item{groups}{a vector with the length of \code{n.groups}, storing the test unit group name.}
\item{group.idx.start}{the start position of each test unit group in group.file. The length of group.idx.start equals to \code{n.groups}.}
\item{group.idx.end}{the end position of each test unit group in group.file. The length of group.idx.start equals to \code{n.groups}..}
\item{scores}{a list with length of \code{n.cohort}. Each element is a martrix with row numbers equal to the total SNPs in each cohort and 15 columns defining the test units. The columns are group name, chromosome, position, reference allele, alternative allele, weight, group.idx, idx, sample size, missrate, alternative frequency, SCORE, file number, variant.idx, and flip, respectively.}
\item{auto.filp}{inherited from input argument \code{auto.flip}.}
}
\description{
{Variant Set Mixed Model Association Tests (the burden test, SKAT, SKAT-O, and an efficient hybrid test to combine the burden test and SKAT) for multiple user-defined test units and a null generalized linear mixed model.

Two steps to perform variant set tests. \code{G.pval} takes the returned R object from \code{G.prep} and uses less memory (if the returned R object from \code{G.prep} is saved to an R data file, the R session is terminated, and this R object is loaded into a new R session for running \code{G.pval}), especially when \code{group.file} contains only a subset of variants from \code{geno.file}.}
}
\examples{
\donttest{
library(StocSum)
library(GMMAT)
library(data.table)
data(example)
attach(example)
seed <- 12345
set.seed(seed)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "StocSum")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "StocSum")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix,MAF.range=c(0,0.5), miss.cutoff = 1)
group.file <- system.file("extdata", "SetID.withweights.txt", package = "StocSum")
obj.prep <- G.prep(out.prefix, n.files = 1, group.file = group.file, auto.flip=F)
save(obj.prep,file=paste0(out.prefix,".prepobj.Rdata"))
}
}
\seealso{
\code{glmmkin2randomvec}, \code{G.stat}, \code{G.pval}
}
\author{
Han Chen, Nannan Wang
}
\keyword{set-based}
\keyword{test}
\keyword{variant}
