% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.R
\name{G.pval}
\alias{G.pval}
\title{Variant Set Mixed Model Association Tests (StocSum)}
\usage{
G.pval(
  G.prep.obj,
  MAF.range = c(1e-07, 0.5),
  MAF.weights.beta = c(1, 25),
  miss.cutoff = 1,
  method = "davies",
  tests = "E",
  rho = c(0, 0.1^2, 0.2^2, 0.3^2, 0.4^2, 0.5^2, 0.5, 1),
  use.minor.allele = FALSE
)
}
\arguments{
\item{G.prep.obj}{A class G.prep object, returned by \code{G.prep}.}

\item{MAF.range}{A numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).}

\item{miss.cutoff}{The maximum missing rate allowed for a variant to be included (default = 1, including all variants).}

\item{method}{A method to compute p-values for SKAT-type test statistics (default = "davies"). "davies" represents an exact method that computes a p-value by inverting the characteristic function of the mixture chisq distribution, with an accuracy of 1e-6. When "davies" p-value is less than 1e-5, it defaults to method "kuonen". "kuonen" represents a saddlepoint approximation method that computes the tail probabilities of the mixture chisq distribution. When "kuonen" fails to compute a p-value, it defaults to method "liu". "liu" is a moment-matching approximation method for the mixture chisq distribution.}

\item{tests}{A character vector indicating which tests should be performed ("B" for the burden test, "S" for SKAT, "O" for SKAT-O and "E" for the efficient hybrid test of the burden test and SKAT). The burden test and SKAT are automatically included when performing "O", and the burden test is automatically included when performing "E" (default = "E").}

\item{rho}{A numeric vector defining the search grid used in SKAT-O (see the SKAT-O paper for details). Not used for the burden test, SKAT or the efficient hybrid test of the burden test and SKAT (default = c(0, 0.1^2, 0.2^2, 0.3^2, 0.4^2, 0.5^2, 0.5, 1)).}

\item{use.minor.allele}{A logical switch for whether to use the minor allele (instead of the alt allele) as the coding allele (default = FALSE). It does not change SKAT results, but Burden (as well as SKAT-O and hybrid test to combine the burden test and SKAT) will be affected. Along with the MAF filter, this option is useful for combining rare mutations, assuming rare allele effects are in the same direction. Use with caution, as major/minor alleles may flip in different cohorts. In that case, minor allele will be determined based on the allele frequency in the combined samples.}

\item{MAF.weight.beta}{A numeric vector of length 2 defining the beta probability density function parameters on the minor allele frequencies. This internal minor allele frequency weight is multiplied by the external weight given by the group.file. To turn off internal minor allele frequency weight and only use the external weight given by the group.file, use c(1, 1) to assign flat weights (default = c(1, 25)). Applied to the combined samples.}
}
\value{
\code{G.pval} returns a data frame with the following components:
\item{groups}{name of the test unit group.}
\item{n.variants}{number of variants in the test unit group that pass the missing rate and allele frequency filters.}
\item{B.score}{burden test score statistic (avaiable if test is Burden, SKATO, or SMMAT).}
\item{B.var}{variance of burden test score statistic (avaiable if test is Burden, SKATO, or SMMAT).}
\item{B.pval}{burden test p-value (avaiable if test is Burden, SKATO, or SMMAT).}
\item{S.pval}{SKAT p-value (avaiable if test is SKATO or SKAT).}
\item{O.pval}{SKAT-O p-value (avaiable if test is SKATO).}
\item{E.pval}{efficient hybrid test of the burden test and SKAT p-value (avaiable if test is SMMAT).}
}
\description{

}
\examples{
\donttest{
library(StocSum)
library(GMMAT)
library(data.table)
data(example)
attach(example)
seed <- 12345
set.seed(seed)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "StocSum")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "StocSum")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix,MAF.range=c(0,0.5), miss.cutoff = 1)
group.file <- system.file("extdata", "SetID.withweights.txt", package = "StocSum")
obj.prep <- G.prep(out.prefix, n.files = 1, group.file = group.file, auto.flip=F)
save(obj.prep,file=paste0(out.prefix,".prepobj.Rdata"))
obj.prep <- get(load(paste0(out.prefix,".prepobj.Rdata")))
out <- G.pval(obj.prep, MAF.range = c(0,0.5), miss.cutoff = 1, method = "davies")
print(out)
unink(paste0(out.prefix,"prepobj.Rdata"))
}
}
\seealso{
\code{glmmkin2randomvec}, \code{G.stat}, \code{G.prep}
}
\author{
Han Chen, Nannan Wang
}
\keyword{set-based}
\keyword{test}
\keyword{variant}
