% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.LDSC.R, R/StocSum.R
\name{G.stat}
\alias{G.stat}
\title{Calculate summary statistics and stochastic statistics}
\usage{
G.stat(
  null.obj,
  geno.file,
  meta.file.prefix,
  MAF.range = c(1e-07, 0.5),
  miss.cutoff = 1,
  missing.method = "impute2mean",
  nperbatch = 10000,
  ncores = 1
)

G.stat(
  null.obj,
  geno.file,
  meta.file.prefix,
  MAF.range = c(1e-07, 0.5),
  miss.cutoff = 1,
  missing.method = "impute2mean",
  nperbatch = 10000,
  ncores = 1
)
}
\arguments{
\item{null.obj}{A class glmmkin.randomvec object, returned by generating random vectors using \code{glmmkin2randomvec}.}

\item{geno.file}{The .gds file name or an object of class SeqVarGDSClass for the full genotypes. The \code{sample.id} in \code{geno.file} should overlap \code{id_include} in \code{null.obj}. It is recommended that \code{sample.id} in \code{geno.file} include the full samples (at least all samples as specified in \code{id_include} of \code{null.obj}). It is not necessary for the user to take a subset of \code{geno.file} before running the analysis. If \code{geno.file} is an object of class SeqVarGDSClass, the .gds file will be closed upon successful completion of the function.}

\item{meta.file.prefix}{Prefix of intermediate files (*.sample.1 and *.resample.1) required in \code{G.pval}.}

\item{MAF.range}{A numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).}

\item{miss.cutoff}{The maximum missing rate allowed for a variant to be included (default = 1, including all variants).}

\item{missing.method}{Method of handling missing genotypes. Either "impute2mean" or "impute2zero" (default = "impute2mean").}

\item{nperbatch}{An integer for how many SNPs to be included in a batch (default = 10000). The computational time can increase dramatically if this value is either small or large. The optimal value for best performance depends on the user’s system.}

\item{ncores}{A positive integer indicating the number of cores to be used in parallel computing (default = 1).}
}
\value{
NULL. \code{G.stat} will store the summary statistics and the stochastic statistics in two files with the prefix specified by user.

NULL. \code{G.stat} will store the summary statistics and the stochastic statistics in two files with the prefix specified by user.
}
\description{
Calculate summary statistics and stochastic statistics.

Calculate summary statistics and stochastic statistics.
}
\examples{
\donttest{
library(StocSum)
library(GMMAT)
library(data.table)
data(example)
attach(example)
seed <- 12345
set.seed(seed)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "StocSum")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "StocSum")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix, MAF.range=c(0,0.5), miss.cutoff = 1)
}
\donttest{
library(StocSum)
library(GMMAT)
library(data.table)
data(example)
attach(example)
seed <- 12345
set.seed(seed)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "StocSum")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "StocSum")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix, MAF.range=c(0,0.5), miss.cutoff = 1)
}
}
\seealso{
\code{glmmkin2randomvec}, \code{G.pval}

\code{glmmkin2randomvec}, \code{G.pval}
}
\author{
Han Chen, Nannan Wang
}
\keyword{statistics}
\keyword{summary}
