% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.GE.R
\name{GE.stat}
\alias{GE.stat}
\title{Calculate stochastic statistics for gene-environment tests}
\usage{
GE.stat(
  null.obj,
  interaction,
  interaction.covariates = NULL,
  geno.file,
  meta.file.prefix,
  MAF.range = c(1e-07, 0.5),
  miss.cutoff = 1,
  missing.method = "impute2mean",
  nperbatch = 10000,
  ncores = 1
)
}
\arguments{
\item{null.obj}{a class glmmkin.randomvec object, returned by generating random vectors using \code{glmmkin2randomvec}.}

\item{interaction}{a numeric or a character vector indicating the environmental factors. If a numberic vector, it represents which indices in the order of covariates are the environmental factors; if a character vector, it represents the variable names of the environmental factors.}

\item{interaction.covariates}{a numeric or a character vector indicating the interaction covariates. If a numeric vector, it represents which indices in the order of covariates are the interaction covariates; if a character vector, it represents the variable names of the interaction covariates.}

\item{geno.file}{the .gds file name or an object of class SeqVarGDSClass for the full genotypes. The \code{sample.id} in \code{geno.file} should overlap \code{id_include} in \code{null.obj}. It is recommended that \code{sample.id} in \code{geno.file} include the full samples (at least all samples as specified in \code{id_include} of \code{null.obj}). It is not necessary for the user to take a subset of \code{geno.file} before running the analysis. If \code{geno.file} is an object of class SeqVarGDSClass, the .gds file will be closed upon successful completion of the function.}

\item{meta.file.prefix}{prefix of intermediate files (*.sample.1 and *.resample.1).}

\item{MAF.range}{a numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).}

\item{missing.method}{method of handling missing genotypes. Either "impute2mean" or "impute2zero" (default = "impute2mean").}

\item{nperbatch}{an integer for how many SNPs to be included in a batch (default = 10000). The computational time can increase dramatically if this value is either small or large. The optimal value for best performance depends on the user’s system.}

\item{ncores}{a positive integer indicating the number of cores to be used in parallel computing (default = 1).}

\item{missing.cutoff}{the maximum missing rate allowed for a variant to be included (default = 1, including all variants).}
}
\value{
NULL. \code{GE.stat} will store the summary statistics and the stochastic statistics in two files with the prefix specified by the user.
}
\description{
Calculate summary statistics and stochastic statistics in the context of gene-environment tests.
}
\examples{
\donttest{
library(GMMAT)
data(example)
attach(example)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "GMMAT")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test.GE"
gdsfile <- system.file("extdata", "geno.gds", package = "GMMAT")
GE.stat(obj, interaction='sex', geno.file = gdsfile, meta.file.prefix = out.prefix)
}
}
\seealso{
\code{glmmkin2randomvec}, \code{GE.pval}
}
\author{
Han Chen, Nannan Wang
}
\keyword{gene-environment}
\keyword{statistics,}
\keyword{summary}
\keyword{tests}
