% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.LDSC.R
\name{LDSC.win}
\alias{LDSC.win}
\title{Calculate LDscore}
\usage{
LDSC.win(
  meta.files.prefix,
  n.files = 1,
  N.randomvec = 1000,
  MAF.range = c(0.001, 0.5),
  miss.cutoff = 1,
  wind.b = 1e+06,
  use.minor.allele = FALSE,
  auto.flip = FALSE,
  nperbatch = 10000,
  ncores = 1
)
}
\arguments{
\item{n.files}{An integer indicating how many sets of intermediate files (.score.* and .var.*), usually as the result of multi-threading in creating the intermediate files (default = 1)).}

\item{N.randomvec}{The number of replicates to simulate the random vectors in \code{StocSum.LDSC.glmmkin2randomvec} (default = 1000).}

\item{MAF.range}{A numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).}

\item{miss.cutoff}{The maximum missing rate allowed for a variant to be included (default = 1, including all variants).}

\item{wind.b}{A positive integer define the window size in bases (default = 1000000).}

\item{use.minor.allele}{A logical switch for whether to use the minor allele (instead of the alt allele) as the coding allele (default = FALSE). It does not change SKAT results, but Burden (as well as SKAT-O and hybrid test to combine the burden test and SKAT) will be affected. Along with the MAF filter, this option is useful for combining rare mutations, assuming rare allele effects are in the same direction. Use with caution, as major/minor alleles may flip in different cohorts. In that case, minor allele will be determined based on the allele frequency in the combined samples.}

\item{auto.flip}{A logical switch for whether to enable automatic allele flipping if a variant with alleles ref/alt is not found at a position, but a variant at the same position with alleles alt/ref is found (default = FALSE). Use with caution for whole genome sequence data, as both ref/alt and alt/ref variants at the same position are not uncommon, and they are likely two different variants, rather than allele flipping.}

\item{nperbatch}{An integer for how many SNPs to be included in a batch (default = 10000). The computational time can increase dramatically if this value is either small or large. The optimal value for best performance depends on the user’s system.}

\item{ncores}{A positive integer indicating the number of cores to be used in parallel computing (default = 1).}

\item{meta.file.prefix}{A character for prefix of intermediate files (*.sample.* and *.resample.*).}
}
\value{
\code{LDSC.win} returns a data frame with the following components:
\item{chr}{chromosome name.}
\item{pos}{the genome location of SNP.}
\item{N}{total sample size.}
\item{altfreq}{alternative allele frequency.}
\item{LDscore}{LD score for each SNP.}
}
\description{
Use summary statistics and stochastic statistics from G.stat to estimate LDscore.
}
\examples{
\donttest{
library(StocSum)
library(GMMAT)
library(data.table)
data(example)
attach(example)
seed <- 12345
set.seed(seed)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "StocSum")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
obj <- LDSC.glmmkin2randomvec(nullmod)
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "StocSum")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix, MAF.range=c(0,0.5), miss.cutoff = 1)
out<-LDSC.win(out.prefix, use.minor.allele = FALSE, auto.flip = FALSE, wind.b = 1000000, nperbatch = 10000)
}
}
\seealso{
\code{LDSC.glmmkin2randomvec}, \code{G.stat}
}
\author{
Han Chen, Nannan Wang
}
\keyword{LD}
\keyword{Score}
