% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.R
\name{svt.meta}
\alias{svt.meta}
\title{Performing meta-analysis for GLMM bashed socre test results}
\usage{
svt.meta(
  meta.files.prefix,
  n.files = rep(1, length(meta.files.prefix)),
  outfile.prefix,
  MAF.range = c(1e-07, 0.5),
  miss.cutoff = 1,
  auto.flip = FALSE,
  nperbatch = 10000
)
}
\arguments{
\item{n.files}{an integer vector with the same length as meta.files.prefix, indicating how many sets of intermediate files (.score.* and .var.*) are expected from each cohort, usually as the result of multi-threading in creating the intermediate files (default = rep(1, length(meta.files.prefix))).}

\item{outfile.prefix}{a character vector for prefix of intermediate files (*.sample.* and *.resample.*) which stores summary statistics and stochatic statistics of meta-analysis.}

\item{MAF.range}{a numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).}

\item{auto.flip}{a logical switch for whether to enable automatic allele flipping if a variant with alleles ref/alt is not found at a position, but a variant at the same position with alleles alt/ref is found (default = FALSE). Use with caution for whole genome sequence data, as both ref/alt and alt/ref variants at the same position are not uncommon, and they are likely two different variants, rather than allele flipping.}

\item{nperbatch}{an integer for how many SNPs to be included in a batch (default = 10000). The computational time can increase dramatically if this value is either small or large. The optimal value for best performance depends on the user’s system.}

\item{meta.file.prefix}{a character vector for prefix of intermediate files (*.sample.* and *.resample.*) which stores summary statistics and stochastic statistics calculated from \code{G.stat}.}

\item{missing.cutoff}{the maximum missing rate allowed for a variant to be included (default = 1, including all variants).}
}
\value{
NULL. \code{svt.meta} will store the summary statistics and the stochastic in two files with the prefix specified by the user.
}
\description{
Use summary statistics and stochastic statistics from \code{G.stat} to peform meta-analysis.
}
\examples{
\donttest{
library(GMMAT)
data(example)
attach(example)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "GMMAT")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "GMMAT")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix,MAF.range=c(0,0.5), miss.cutoff = 1)
NOT DONE!!!!!!!!!!!!!!!!
}
}
\seealso{
\code{glmmkin2randomvec}, \code{G.stat}, \code{G.pval}
}
\author{
Han Chen, Nannan Wang
}
\keyword{meta-analysis}
