% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.R
\name{svt.pval}
\alias{svt.pval}
\title{Performing GLMM based score tests}
\usage{
svt.pval(
  meta.files.prefix,
  n.files = rep(1, length(meta.files.prefix)),
  MAF.range = c(1e-07, 0.5),
  miss.cutoff = 1,
  auto.flip = FALSE,
  nperbatch = 10000
)
}
\arguments{
\item{n.files}{an integer vector with the same length as meta.files.prefix, indicating how many sets of intermediate files (.score.* and .var.*) are expected from each cohort, usually as the result of multi-threading in creating the intermediate files (default = rep(1, length(meta.files.prefix))).}

\item{MAF.range}{a numeric vector of length 2 defining the minimum and maximum minor allele frequencies of variants that should be included in the analysis (default = c(1e-7, 0.5)).}

\item{auto.flip}{a logical switch for whether to enable automatic allele flipping if a variant with alleles ref/alt is not found at a position, but a variant at the same position with alleles alt/ref is found (default = FALSE). Use with caution for whole genome sequence data, as both ref/alt and alt/ref variants at the same position are not uncommon, and they are likely two different variants, rather than allele flipping.}

\item{nperbatch}{an integer for how many SNPs should be tested in a batch (default = 10000). The computational time can increase dramatically if this value is either small or large. The optimal value for best performance depends on the user’s system.}

\item{meta.file.prefix}{a character vector for prefix of intermediate files (*.sample.* and *.resample.*).}

\item{MAF.weight.beta}{a numeric vector of length 2 defining the beta probability density function parameters on the minor allele frequencies. This internal minor allele frequency weight is multiplied by the external weight given by the group.file. To turn off internal minor allele frequency weight and only use the external weight given by the group.file, use c(1, 1) to assign flat weights (default = c(1, 25)). Applied to the combined samples.}

\item{missing.cutoff}{the maximum missing rate allowed for a variant to be included (default = 1, including all variants).}
}
\value{
\code{svt.pval} returns a data frame with the following components:
\item{SNP}{SNP name.}
\item{chr}{chromosome name.}
\item{pos}{the genome location of SNP.}
\item{ref}{allele of reference.}
\item{alt}{alternative allele.}
\item{N}{total sample size.}
\item{missrate}{missing rate of variants.}
\item{altfreq}{alternative allele frequency.}
\item{SCORE}{the summary score of the alternaive allele.}
\item{VAL}{the variance of the summary score.}
\item{PVAL}{the p-value of the suammry score.}
}
\description{
Perform score tests for association with genotypes in a GDS file .gds, with summary statistics calculated from \code{glmmkin2randomvec}.
}
\examples{
\donttest{
library(GMMAT)
data(example)
attach(example)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "GMMAT")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
if(!is.null(nullmod$P)){
  obj <- glmmkin2randomvec(nullmod)
}else{
  kinship.chol <- chol(GRM)
  obj<-glmmkin2randomvec(nullmod, Z = list(t(kinship.chol)))
}
out.prefix <- "test"
gdsfile <- system.file("extdata", "geno.gds", package = "GMMAT")
G.stat(obj, geno.file = gdsfile, meta.file.prefix = out.prefix,MAF.range=c(0,0.5), miss.cutoff = 1)
out <- svt.pval(out.prefix, MAF.range=c(1e-7, 0.5), miss.cutoff = 1, auto.flip=F)
}
}
\seealso{
\code{glmmkin2randomvec}, \code{G.stat}
}
\author{
Han Chen, Nannan Wang
}
\keyword{generalized}
\keyword{linear}
\keyword{mixed}
\keyword{model}
\keyword{score}
\keyword{test,}
