% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocSum.LDSC.R
\name{LDSC.glmmkin2randomvec}
\alias{LDSC.glmmkin2randomvec}
\title{Create random vectors for a glmmkin object for LD Score regresion}
\usage{
LDSC.glmmkin2randomvec(
  obj,
  Z = NULL,
  N.randomvec = 1000,
  group.idx = NULL,
  cluster.idx = NULL,
  robust = FALSE
)
}
\arguments{
\item{obj}{The glmmkin object.}

\item{Z}{NULL. %% A list of design matrices for the random effects. The length must match the number of variance components.}

\item{N.randomvec}{The number of random vectors to generate (default = 1000).}

\item{group.idx}{NULL. %%A length N index vector showing which observation belongs to which variance group, for heteroscedastic linear mixed models (default = NULL for homoscedastic linear mixed models).}

\item{cluster.idx}{NULL. %% A length N index vector showing which observation belongs to which cluster (default = NULL for no clusters).}

\item{robust}{FALSE. %% A logical switch: whether robust variance should be used (default = FALSE).}
}
\value{

}
\description{
Generate random vectors from multivariate normal distribution with mean 0 and covariance matrix P, in which \eqn P = I-1inv((t(1)1))t(1).
}
\examples{
\donttest{
library(GMMAT)
data(example)
attach(example)
GRM.file <- system.file("extdata", "GRM.txt.bz2", package = "GMMAT")
GRM <- as.matrix(read.table(GRM.file, check.names = FALSE))
nullmod <- glmmkin(disease ~ age + sex, data = pheno, kins = GRM, id = "id", family = binomial(link = "logit"))
obj <- LDSC.glmmkin2randomvec(nullmod)
}
}
\author{
Han Chen, Nannan Wang
}
\keyword{random}
\keyword{vector}
